* Various PL_unify_term() issues

+++ src/pl-umap.c
@@ -1629,7 +1629,7 @@ uflagsW(int code)
   if ( cp < UNICODE_MAP_SIZE )
   { const char *s = uflags_map[cp];
     if ( s < (const char *)256 )
-      return (int)s;
+      return (int)(long)s;
     return s[code&0xff];
   }
   return 0;
